分页查询在线用户修复

+++ src/main/java/net/chenlin/dp/modules/biz/bussiness/entity/YyMOnlineDayEntity.java
@@ -32,7 +32,7 @@ public class YyMOnlineDayEntity implements Serializable {
 	private Date createTime;
 	
 	@ApiModelProperty(value = "每日在线用户数量")
-	private Double mOnlineNumber;
+	private Long mOnlineNumber;
